More Changes

+++ Comex/Comex.Modelos/Produtos/Produto.cs
@@ -10,7 +10,7 @@ public class Produto
     public string Nome { get; set; }
     public string Descricao { get; set; }
     public decimal Preco { get; set; }
-    public int Estoque { get; set; }
+    public decimal Estoque { get; set; }
     public Categoria Categoria { get; set; }
     protected string Atributos { get; set; }
